Improve model building
Using only Estonian loans as Finnish and Spanish loans
behave differently and there is not sufficient data for
creating a reliable model for those. Additionally, using
all loans that are at least 2 years old as it allows making
use of more data and after 2 years 95% of defaults have
already happened.